build: 静态分析零警告 + 配置异常安全修复 + ADR 索引升级
详细变更:

src/（实际代码修复）
- config_loader.cpp/hpp：5 个 helper 函数增加 try/catch(YAML::Exception) → CONFIG_INVALID_VALUE，
  修复 WrongTypeReturnsErr 测试（之前 .as<T>() 抛 TypedBadConversion 触发 noexcept → std::terminate）；
  load_from_file 改 static（无 this 访问）
- psk.cpp/psk_authenticator.cpp：未初始化变量 tmp 修复
- topology.cpp/topology.hpp：commit() 加 NOLINT(rvalue-reference-not-moved)
- buffer.hpp：DynamicRingBuffer 字段重排消除 144 字节 padding
- ability_a/transport/udp_socket.cpp：ret 初始化
- ability_c/executor/process_executor.cpp：n 初始化
- ability_c/nodes/nodes.cpp：空 catch 加 (void)0 + NOLINTNEXTLINE
- core/log/logger.cpp：fprintf/snprintf 加 NOLINT(cert-err33-c)
- ability_b/transport/inproc_channel.cpp：析构虚调用改 NOLINTNEXTLINE
- sdk/udaf_c.cpp：delete 加 NOLINTNEXTLINE(owning-memory)
- crypto/pki.cpp：operator() 加 NOLINTNEXTLINE(owning-memory)
- observability.hpp/cpp：Tracer::begin 改 static
- platform/fs/wal.hpp/cpp：WalEntry 加 NOLINT（kSchemaVersion static constexpr 误报）；
  新增 scan_entries_for_max_seq 实现（read_file_header 调用）

cmake/StaticAnalyzers.cmake
- 新增 -clang-analyzer-optin.cplusplus.VirtualCall（析构中显式调用虚函数用于清理）
- 新增 -readability-convert-member-functions-to-static（保留 const + 实例语义）
- 完整注释说明每个排除项的理由（PIMPL / OpenSSL / 序列化 / 虚调用清理）

tests/（测试同步）
- test_observability.cpp：Tracer 调用改静态 begin/end/inject/extract
- test_ability_a / test_audit / test_buffer / test_crypto / test_logger / test_integration /
  test_cli / test_ability_c / test_c_api / test_wal：覆盖率补充 + NOLINT 同步

docs/（设计文档一致性）
- 02-architecture.md v2.9 → v2.10：附录 B ADR 索引新增"状态"+"关键决策"列；
  ADR-001~010 由"提议（待评审）"批量更新为"已批准"（附批准日期 2026-09-01）
- 03-detailed-design.md v2.2 → v2.3：前置引用同步（02 v2.10）
- 04-module-design.md v0.8 → v0.9：前置引用同步（02 v2.10 + 03 v2.3）
- 05-test-plan.md v0.8 → v0.9：前置引用同步（02 v2.10 + 03 v2.3 + 04 v0.9），§5.14 风险表版本号引用更新
- ADR-001~010：状态"提议"→"已批准"+ 批准日期

CHANGELOG.md：v0.3.14 条目（静态分析零警告 + 异常安全修复 + ADR 索引升级）

指标
- 测试：446/446 通过（+1 修复 WrongTypeReturnsErr）
- 静态分析：clang-tidy 全检查列表 0 actionable warnings
- ADR 状态：11/11 已批准
- 文档一致性：02/03/04/05 四方版本号引用全部对齐

+++ src/audit/audit.hpp
@@ -21,7 +21,7 @@
 namespace udaf::audit {
 
 // 19 项 ActionType（ADR-006 §2.1 + v1.7 扩展 6 项）
-enum class ActionType : std::uint16_t {
+enum class ActionType : std::uint8_t {
     NodeRegister            = 1,
     NodeUnregister          = 2,
     NodeHeartbeat           = 3,
